+++ examples/sa-token-example/src/sea_orm_storage.rs
@@ -5,7 +5,9 @@
 
 use async_trait::async_trait;
 use summer::plugin::service::Service;
-use summer_sa_token::sa_token_adapter::storage::{SaStorage, StorageError, StorageResult};
+use summer_sa_token::sa_token_plugin_axum::sa_token_adapter::storage::{
+    SaStorage, StorageError, StorageResult,
+};
 use sea_orm::entity::prelude::*;
 use sea_orm::{ColumnTrait, EntityTrait, QueryFilter, Set};
 use summer_sea_orm::DbConn;